test(harness): long-run differential tests + f64 comparison tightenings
Stage 6.4b diagnostic infrastructure. Extends the existing
test_analysis_run_matches_c_reference with two new tests that prove
the analysis module is bit-exact on longer and real-music content:

- test_analysis_run_matches_c_reference_long: 500 frames of synthetic
  3-tone + pseudorandom noise, compares full TonalityAnalysisState
  after EVERY frame (not just at the end). Passes bit-exact.
- test_analysis_run_matches_c_reference_real_music: 300 frames loaded
  from tests/vectors/music_48k_stereo.wav via downmix_int (matching
  opus_encode). Passes bit-exact.

Together these eliminate the hypothesis that the remaining ~5% byte
divergence in full-pipeline music encode comes from a latent analysis
module bug not caught by the original 30-frame sine test. The
analysis module produces bit-identical output to C for the exact
audio content the encoder sees.

Also tightens two f32/f64 comparison sites in ropus/src/celt/encoder.rs
where C uses bare-double literals that were being compared at f32 in
Rust:
- analysis.tonality > 0.3  (line 2466, alloc_trim_analysis)
- analysis.activity  < 0.4 (line 2121, compute_vbr)

0.3 and 0.4 are not exactly representable in f32 (round up / round
down respectively), so a threshold comparison can flip between f32
and f64 precision. Promoted both to f64 to match C semantics. These
changes are correct but did NOT reduce the DIFFER count on
music_48k_stereo.wav — the content doesn't hit these borderline
cases. Kept as defensive transcription fidelity.

Stage 6.4b hit the bounded 2-round diagnostic cap without reducing
the 457/104 DIFFER floor (music/speech). The remaining drift is
not in analysis output, nor in the 9 analysis consumption wiring
sites. Agent's hypothesis: sub-ULP drift in long-running CELT
encoder state (stereo_saving, tonal_average, hf_average,
hybrid_stereo_width_Q14, etc.). Localizing this would require a
full-struct FFI byte-compare of CELTEncoder between C and Rust —
out of the bounded budget.

Tier-1 byte-exact not achieved for music/speech. Next decision point
pending: attempt another round targeting CELTEncoder state, or drop
to tier-2 (decision agreement + bitrate stability) per HLD v2.

+++ ropus/src/celt/encoder.rs
@@ -2115,7 +2115,10 @@ fn compute_vbr(
     //      target -= (opus_int32)((coded_bins<<BITRES)*(.4f-analysis->activity));
     // The f32 product casts to i32 with truncation toward zero, then the
     // `target -= …` is an integer subtraction.
-    if analysis.valid != 0 && analysis.activity < 0.4_f32 {
+    // Note: `.4` in C is a bare double literal; `float < double` promotes
+    // to f64 for the comparison. The inner `(.4f-analysis->activity)` uses
+    // a single-precision `.4f` literal — kept in f32 for that arithmetic.
+    if analysis.valid != 0 && (analysis.activity as f64) < 0.4_f64 {
         let reduction = ((coded_bins << BITRES) as f32 * (0.4_f32 - analysis.activity)) as i32;
         target -= reduction;
     }
@@ -2458,8 +2461,12 @@ fn celt_encode_core(
     //               && (pitch_index > 1.26*prev || pitch_index < .79*prev);
     // The analysis gate prevents spurious pitch-change decisions when the
     // tonality analyzer says the current frame isn't actually tonal.
+    // Note: `.3` in C is a bare double literal, so `tonality > .3` promotes
+    // the f32 to f64 before comparison. Replicate that precisely — an f32
+    // comparison against `0.3_f32` uses a slightly higher threshold (since
+    // `0.3` rounds up in f32) and can flip the branch on borderline values.
     pitch_change = if (gain1 > qconst16(0.4, 15) || st.prefilter_gain > qconst16(0.4, 15))
-        && (st.analysis.valid == 0 || st.analysis.tonality > 0.3_f32)
+        && (st.analysis.valid == 0 || (st.analysis.tonality as f64) > 0.3_f64)
         && (pitch_index as f64 > 1.26 * st.prefilter_period as f64
             || (pitch_index as f64) < 0.79 * st.prefilter_period as f64)
     {
